Reduced metrics to SEC fields and simplified icon
Removed all calculated ratios since they aren't available from the SEC and will need to be added after the other metrics are available.

Switched logo to simple bar-and-line chart (was temporarily a complex image with CI shadow). Note the logos are fuzzy and will need to be recreated in Sketch later -- this is just a placeholder.