Move DB using Swift, create Views directory for UIView subclasses, rename ScrollChartController to ScrollChartView
ScrollChartView has a fair amount of rendering logic which is why it was previously called ScrollChartController. But so does WKWebView and it is a view.

Rewrote DB helper as DBUpdater.swift -- additional code will be added to make an API request to find stocks to add (for IPOs), delete (for delistings), and clear cached data (for stock splits).