Added nullability and collection type annotations for Swift migration
Since much of the project was written in 2012 there are still some remnants from before Automatic Reference Counting. This commit removes some dealloc methods which only cleared the (RootViewController *) delegate. Since the RVC remains in memory as long as the app is running, it wasn't necessary to reduce the references to it.