Swift WatchlistViewController replaces ObjC RootViewController
While WatchlistViewController is the first full-screen controller shown, technically window.rootViewController = tabBarController
The tabBarController contains a UINavigationController which contains the WatchlistViewController.

One bug was also fixed where a monthly chart comparison of a recent IPO and an established company would show offset dates for the recent IPO. For now, this was fixed by limiting the maximum date range to the shortest date range of the companies being compared.