Swift StockData rewrite
The prior Objective-C version used a concurrent queue with barrier blocks to allow multi-threaded background rendering with the barrier blocks used to serialize access to shared data. To speed up the rewrite this Swift version also uses a concurrent queue. Once ScrollChartView is rewritten to Swift I intend to rewrite it to use the new Swift concurrency features async and await.