WebView to browse chartinsight.com with button to add to watchlist
To keep the main watchlist screen simple, a middle tab was added that allows exploring the stocks on chartinsight.com by sector, insider buying, and 13F ownership. On a stock detail page, a lower toolbar button appears to either add the stock to the watchlist or switch to view an existing comparison with that stock.

Also moved Metrics data into a singleton and out of AppDelegate.