Get stock changes (splits, additions, ticker changes) from API
Because historical price data is cached on the device, it is necessary to clear out the cached history for a stock when a stock split occurs. The API also indicates when stocks have been delisted/acquired and they will be removed from the stock search and any existing comparisons. Also switched stock search to avoid the DBActor isolation since the actor isolation delayed search results while historical data was being saved to the history table. That wasn't necessary since the stock table can be opened read-only for full-text searches.