Updated DB, ChartElements is a struct for easy copying
Updated the database and added a few ETFs (SPY, DIA, IWM, QQQ, etc.)

Completed the async version of magnifyBar(xPress:yPress:) to show a magnified section of the chart when the user long presses.

Renamed tmp to chartElements on stockActor since there is now only one ChartElements property and a copy is returned via copyChartElements().

The StockChangeService has been finished and it can be used to add additional stocks or ETFs later.

Now that the app has been rewritten to Swift, the copy behavior of structs allows avoiding the copy(with zone:) requirement needed to pass ChartElements to Objective-C code. BarData is still a class because it has a static method which calculates the moving averages on an array of BarData where accessing the references is important.