DBActor returns updated comparison list after adding or removing a stock
Pushing new changes to the tableView instead of having it pull them from DBActor avoids data races and will be faster since it avoids an extra db open/close and task dispatch to DBActor.

Also improved the logic for updating the stockActor list when adding or removing a stock or the entire comparison.