Update stockActor.stock when user changes chart options
To ensure that chart options are persisted to the DB and to the list of all stock comparisons, a new list of stocks is reloaded after updating options. When a stock is removed, the stockActor (and any network sessions) for that comparisonStockId is invalidated.

Also added a URL input field to the WebViewController so the user can see the current URL and enter a new one for ChartInsight.com magic link authentication.