Stock is now a struct, added public, private and final annotations
Even with the prior Stock class, it was necessary to update the stock property of ChartRenderer after each chart options change to a stock object. So it made sense to switch Stock from a class to a struct to make the value semantics clear. To update the comparisonStockId after a stock is added to a comparison, the insertedComparisonStockId is returned by the DBActor.

HistoricalDataService is the new name for DataFetcher.
FundamentalService is the new name for FundamentalFetcher.

Most of the changes were explicitly marking functions and variables "public" and classes "final".