ScrollChartViewModel.didUpdate() closure now passes ChartElements to View
Previously ScrollChartView.renderCharts() had to await a method on ScrollChartViewModel to get a copy of the ChartElements. While ScollChartView.renderCharts() still needs to copy some values from the ViewModel to the ChartRenderer, it doesn't need to request any StockActor-isolated values and thus doesn't need to be async.

ChartStyleControl.delegate is now a weak var to avoid retain cycles.
WatchlistViewController.popOverNav is now optional so the popOver navigation controller and all child controllers are released after the user dismisses them.