Use autogenerated memberwise initializer for ChartElements, update comments
ChartElements was a class during the rewrite from Objective-C but is now a struct with an compiler-generated memberwise initializer. While a stock instance must always be provided to the initializer, default values for the other properties allow ChartElements(stock:) to use the compiler-generated initializer.

Also updated comments to clarify when processing is done by ScrollChartViewModel, ScrollChartView, or ChartRenderer.