Async fetch* methods for HistoricalDataService and FundamentalService
Callers of those methods already had to use await to serialize requests to the StockActor. Replacing the internal Task with an external async signature reduces the nesting within each method.

Also changed the delegate for DBActor to a protocol so the tests can act as the delegate and delete cached price history by passing in a fake stock split.

Added a background to the app-core-objects.svg file so it is readable in night mode.

Moved delegate?.requestStarted() method that starts the progress indicator within StockActor.fetchPriceAndFundamentals().