Unit tests for ScrollChartViewModel
To simplify testing of ScrollChartViewModel.scaleChart(newScale:pxShift:), the update to the xFactor property was moved out of the Task.

An XCTestExpectation is used to verify that the didUpdate(_) closure is called after data is fetched.

BarData was moved to its own file (it had been in the pending changes to ScrollChartViewModel and the changes to ScrollChartViewModel.swift were held back from the last commit.)